Extract run blocks into functions
